move simulators to their own folder
